Refactored the project. Created project structure and added methods to Board.
