--- src/main/java/PlayerInput.java
@@ -1,45 +1,0 @@
-import org.w3c.dom.ls.LSOutput;
-
-import java.security.KeyStore;
-import java.util.Scanner;
-
-public class PlayerInput {
-
-    //TODO Need a while statement for game status to determine if game is done or not.
-    // playerInput asks for player 'x' or 'o' if they are going
-    // then after the input we will reprint the tic tac toe game-board
-
-    //fields
-    boolean playerX = true;
-    boolean playerO;
-
-//    Scanner scan = new Scanner(System.in);
-//    System.out.println("Player 1 please enter your first move by row and columnusing numbers 1-9");
-
-//        while(finish.equals)("")){
-//    if(playerX){
-//        System.out.println("Player x turn");
-//     char[]
-//    }else{
-//        System.out.println("Player o turn");
-//    }
-//        char[][] gameBoard; // need to make board a separate class
-//        Board.printGameBoard(gameBoard);
-
-
-    //TODO check to see if zone is occupied
-
-    //TODO plug input on the game
-    //getPlayerInput(playerx, gameboard);
-
-    /*
-    here we will need an if statement to check to see if the game is over and print the board
-    Also, need to check the board to see if cat game or not.
-     */
-//        if winner_winner(three_in_row){
-//            game_over = true;
-//            System.out.println("Winner Winner");
-//        }
-
-//    }
-}
