+++ src/main/java/com/tictactoe/winconditions/WinConditions.java
@@ -1,4 +1,4 @@
-
+package com.tictactoe.winconditions;
 
 public class WinConditions {
